requirements.txt
pandas >= 1.2.0
numpy >= 1.20.0
scipy >= 1.6.0
websockets>=9.0
websocket-client~=1.3.2